pyproject.toml
[project]
name = "metalgrow"
version = "0.0.1"
description = "AI-powered image upscaler accelerated on Apple Metal (MPS)."
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3",
    "torchvision>=0.18",
    "pillow>=10.0",
    "numpy>=1.26",
    "typer>=0.12",
]

[project.scripts]
metalgrow = "metalgrow.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metalgrow"]

[tool.ruff]
line-length = 100
target-version = "py311"
